Stop the connection reaper thread when pool is thrown away
A ConnectionPool's reaper thread is never stopped. While in production
this won't matter much, it creates a lot of threads during testing.
Also, in reaper_test.rb reapers with a very high frequency are created
but never stopped. These cause unnecessary CPU load and slow down testing.

Since the threads keep running and have a reference to the ConnectionPool
itself, GC can't collect the pool.

The method of shutting down implemented here is possibly racy, but it
won't matter much since the worst case is that the thread gets shutdown
an interval later.

This has been found during testing of JRuby's activerecord-jdbc-adapter.
During a run of the PostgreSQL tests, this commit brings down the live
threads at the end of testing from 185 to 17 (which includes JVM and
JRuby threads).